Move question module to more appropriate folder
